+++ src/yetibot/core/webapp/resolvers.clj
@@ -1,42 +1,61 @@
 (ns yetibot.core.webapp.resolvers
   (:require
-    [yetibot.core.models.users :as users]
-    [yetibot.core.chat :as chat]
-    [yetibot.core.models.channel :as channel]
-    [yetibot.core.webapp.resolvers.stats :as stats]
-    [cuerdas.core :refer [kebab snake]]
-    [yetibot.core.adapters.adapter :as adapter]
-    [yetibot.core.db.history :as history]
-    [yetibot.core.db.alias :as db.alias]
-    [yetibot.core.db.observe :as db.observe]
-    [yetibot.core.db.cron :as db.cron]
-    [yetibot.core.models.karma :as karma]
-    [yetibot.core.handler :refer [handle-unparsed-expr]]
-    [taoensso.timbre :refer [error debug info color-str]]))
+   [yetibot.core.commands.channel :as channel]
+   [yetibot.core.util.format :refer [format-data-structure]]
+   [yetibot.core.adapters.adapter :as adapter]
+   [yetibot.core.models.users :as users]
+   [yetibot.core.chat :as chat]
+   [yetibot.core.webapp.resolvers.stats :as stats]
+   [cuerdas.core :refer [kebab snake]]
+   [yetibot.core.db.history :as history]
+   [yetibot.core.db.alias :as db.alias]
+   [yetibot.core.db.observe :as db.observe]
+   [yetibot.core.db.cron :as db.cron]
+   [yetibot.core.models.karma :as karma]
+   [yetibot.core.handler :refer [handle-unparsed-expr]]
+   [taoensso.timbre :refer [error debug info color-str]]))
 
-(def eval-chat-source {:adapter :graphql :uuid :graphql :room "graphql"})
 (def eval-user {:username "graphql" :name "graphql" :id "graphql"})
+(def eval-channel "graphql")
+
+(comment
+  (binding [chat/*adapter* (adapter/web-adapter)]
+    (let [cs (chat/chat-source eval-channel)
+          expr "channel settings"]
+      (handle-unparsed-expr cs eval-user expr)))
+
+  (binding [chat/*adapter* (adapter/web-adapter)]
+    (let [[f _]
+          (format-data-structure (channel/settings-for-channel eval-channel))]
+      f)))
 
 (defn eval-resolver
   [context {:keys [expr] :as args} value]
   (debug "eval-resolver" args)
-  ;; stub in a fake chat-source and user so commands that depend on these still
-  ;; work
-  (let [{:keys [value error]} (handle-unparsed-expr
-                                eval-chat-source eval-user expr)
-        result (or value error)]
-    ;; normalize to always returning a collection, as required by the graphql
-    ;; schema
-    (if (coll? result)
-      result
-      [result])))
+  ;; bind the web adapter
+  (binding [chat/*adapter* (adapter/web-adapter)]
+    ;; stub in a chat-source and user if the web-adapter is present so commands
+    ;; that depend on these still work
+    (let [cs (and
+              (adapter/web-adapter)
+               (chat/chat-source "graphql"))
+          {:keys [value error]} (handle-unparsed-expr
+                                 cs eval-user expr)
+          result (or value error)
+          [formatted-response _] (format-data-structure result)]
+      ;; normalize to always returning a collection, as required by the graphql
+      ;; schema
+      (if (coll? formatted-response)
+        formatted-response
+        [formatted-response]))))
 
 (comment
-  (handle-unparsed-expr
-    eval-chat-source
-    eval-user
-    "echo hi")
-  )
+  (binding [chat/*adapter* (web-adapter)]
+    (let [cs (chat/chat-source "graphql")]
+      (handle-unparsed-expr
+       cs
+       eval-user
+       "echo hi | raw all"))))
 
 (defn adapters-resolver
   [context {:keys [] :as args} value]
@@ -49,6 +68,15 @@
                     :connection_latency (adapter/connection-latency %)
                     :connection_last_active_timestamp
                     (adapter/connection-last-active-timestamp %)))))
+
+(comment
+
+  (adapter/uuid (last (adapter/active-adapters)))
+  *e
+
+  (->>
+   @adapter/adapters
+   (map adapter/uuid)))
 
 (defn history-resolver
   [context